refactor(backend): extract optical drive service into browsing, burning, and reading modules
